Split tuples
